requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
langchain==1.2.0
langchain-core>=1.2.6
langchain-openai==1.1.6
langgraph==0.0.26
python-dotenv==1.0.1
azure-search-documents==11.4.0
httpx==0.27.0
PyPDF2==3.0.1
azure-core
tortoise-orm==0.20.0
aerich==0.7.2